+++ HISTORY.md
@@ -11,6 +11,7 @@
 - Run peddy on the population variant file, not the individual sample level file if joint calling was done.
 - Add STAR to MultiQC metrics.
 - Throw an error if STAR is run on a genome with alts.
+- Don't run bcbioRNASeq if there is only one sample. Thanks to @kmendler for the suggestion.
 
 ## 1.1.9 (5 December 2019)
 - Fix for get VEP cache.